+++ plugins/content-manager/src/test/java/com/wazuh/contentmanager/transport/TransportTriggerUpdateActionTests.java
@@ -37,6 +37,7 @@ import java.util.concurrent.ExecutorService;
 import com.wazuh.contentmanager.action.MessageStatusResponse;
 import com.wazuh.contentmanager.action.TriggerUpdateRequest;
 import com.wazuh.contentmanager.cti.catalog.index.ConsumersIndex;
+import com.wazuh.contentmanager.cti.catalog.service.EngineContentLoader;
 import com.wazuh.contentmanager.engine.service.EngineService;
 import com.wazuh.contentmanager.jobscheduler.jobs.CatalogSyncJob;
 import com.wazuh.contentmanager.settings.PluginSettings;
@@ -148,7 +149,8 @@ public class TransportTriggerUpdateActionTests extends OpenSearchTestCase {
                         mock(ConsumersIndex.class),
                         mock(Environment.class),
                         threadPool,
-                        mock(EngineService.class));
+                        mock(EngineService.class),
+                        mock(EngineContentLoader.class));
         TransportTriggerUpdateAction realAction =
                 new TransportTriggerUpdateAction(
                         mock(TransportService.class), mock(ActionFilters.class), realCatalogSyncJob);